Remove recipient-specific customization of body using Liquid
SparkPost offers data substitution features that make it unnecessary
to handle this login in-app.

The enduser syntax is eerily similar to Liquid's. The main advantage
of using this method is that instead of sending many emails out
individually from the application, just one request is made to the
SparkPost API with a body and a JSON object with the substitution
data.

This especially makes a difference when sending mails with attachments,
but also helps to manage sent emails better.